+++ src/data/compiled/mod.rs
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-// Copyright (C) 2026 Vallés Puig, Ramon
-
-//! Crate-private JPL planetary ephemeris data sub-modules.
-
-pub(crate) mod jpl;